requirements.txt
openai
pandas
python-docx
snscrape
feedparser